Rename: model 패키지 분리
객체 관련 -> domain, 기능 관련 -> service로 패키지 분리
