+++ src/main/java/model/domain/Car.java
@@ -1,4 +1,4 @@
-package model;
+package model.domain;
 
 public class Car {
